Fix scroll on first render
