+++ packages/ui/react-ui-table/src/components/Table/Table.tsx
@@ -27,6 +27,7 @@ import { TableFooter } from './TableFooter';
 import { TableHead } from './TableHead';
 import { type TableProps } from './props';
 import { groupTh, tableRoot } from '../../theme';
+import { usePinLastRow } from '../hooks/usePinLastRow';
 
 export const Table = <TData extends RowData>(props: TableProps<TData>) => {
   const {
@@ -149,24 +150,7 @@ export const Table = <TData extends RowData>(props: TableProps<TData>) => {
     onDataSelectionChange?.(Object.keys(rowSelection).map((id) => table.getRowModel().rowsById[id].original));
   }, [onDataSelectionChange, rowSelection, table]);
 
-  useEffect(() => {
-    if (!pinLastRow) {
-      return;
-    }
-
-    // Clear row pinning
-    table.resetRowPinning();
-
-    const rows = table.getRowModel().rows;
-    rows[rows.length - 1].pin('bottom');
-    // Scroll to the bottom
-
-    const scrollElement = getScrollElement?.();
-
-    if (scrollElement) {
-      scrollElement.scrollTo({ top: scrollElement.scrollHeight });
-    }
-  }, [pinLastRow, table, data]);
+  usePinLastRow(pinLastRow, table, data, getScrollElement);
 
   // Create additional expansion column if all columns have fixed width.
   const expand = false; // columns.map((column) => column.size).filter(Boolean).length === columns?.length;
